Schematic entry for most parts of project alpha
Design has WiFi module, I2C EEPROM, LDO, LED, SD card, and matching
circuit for antenna.  The antenna isn't included yet as that part
still needs to be created in the library (going with the ACX reference
design part).

Also included solder jumper pads for end user customizations for I2C
EEPROM address, SD card chip select and CC3000 chip select.  These will
be on the back side of the board for easy customization and rewiring.
The user would need to cut the thin trace (10mil) and then resolder to
the pin they want to use (as well as change the software accordingly).

Signed-off-by: fangsuei <fangsuei@fangsuei.com>